+++ packages/docs/components/Examples/linkify/CustomComponentLinkifyEditor/index.js
@@ -2,6 +2,8 @@ import React, { Component } from 'react';
 import { EditorState } from 'draft-js';
 import Editor from '@draft-js-plugins/editor';
 import createLinkifyPlugin from '@draft-js-plugins/linkify';
+import linkifyIt from 'linkify-it';
+import tlds from 'tlds';
 import editorStyles from './editorStyles.module.css';
 
 const linkifyPlugin = createLinkifyPlugin({
@@ -9,6 +11,8 @@ const linkifyPlugin = createLinkifyPlugin({
     // eslint-disable-next-line no-alert, jsx-a11y/anchor-has-content
     return <a {...props} onClick={() => alert('Clicked on Link!')} />;
   },
+  customExtractLinks: (text) =>
+    linkifyIt().tlds(tlds).set({ fuzzyEmail: false }).match(text),
 });
 const plugins = [linkifyPlugin];