+++ packages/linkify/src/Link/Link.tsx
@@ -1,11 +1,7 @@
 import React, { ComponentType, ReactElement, ReactNode } from 'react';
 import clsx from 'clsx';
-import linkifyIt from 'linkify-it';
-import tlds from 'tlds';
 import { LinkifyPluginTheme } from '../theme';
-
-const linkify = linkifyIt();
-linkify.tlds(tlds);
+import { ExtractLinks, extractLinks } from '../utils/extractLinks';
 
 export interface ComponentProps {
   children: ReactNode;
@@ -23,6 +19,7 @@ export interface LinkProps {
   target?: string;
   rel?: string;
   className?: string;
+  customExtractLinks?: ExtractLinks;
 
   // following props are not used
   entityKey?: unknown;
@@ -41,6 +38,7 @@ export default function Link(props: LinkProps): ReactElement {
   const {
     decoratedText = '',
     theme = {} as LinkifyPluginTheme,
+    customExtractLinks = (text) => extractLinks(text),
     target = '_self',
     rel = 'noreferrer noopener',
     className,
@@ -58,7 +56,7 @@ export default function Link(props: LinkProps): ReactElement {
   } = props;
 
   const combinedClassName = clsx(theme?.link, className);
-  const links = linkify.match(decoratedText);
+  const links = customExtractLinks(decoratedText);
   const href = links && links[0] ? links[0].url : '';
 
   const linkProps = {