+++ packages/linkify/src/index.tsx
@@ -3,6 +3,7 @@ import { EditorPlugin } from '@draft-js-plugins/editor';
 import Link, { LinkProps, ComponentProps } from './Link/Link';
 import linkStrategy from './linkStrategy';
 import { defaultTheme, LinkifyPluginTheme } from './theme';
+import { ExtractLinks, extractLinks } from './utils/extractLinks';
 
 export { extractLinks } from './utils/extractLinks';
 
@@ -11,6 +12,11 @@ export interface LinkifyPluginConfig {
   theme?: LinkifyPluginTheme;
   target?: string;
   rel?: string;
+  /**
+   * Custom extract links function that should return Array of index, lastIndex & url.
+   * @param {string} text - Current state of the editor as a plain text.
+   */
+  customExtractLinks?: ExtractLinks;
 }
 
 export default (config: LinkifyPluginConfig = {}): EditorPlugin => {
@@ -26,6 +32,7 @@ export default (config: LinkifyPluginConfig = {}): EditorPlugin => {
     theme = defaultTheme,
     target = '_self',
     rel = 'noreferrer noopener',
+    customExtractLinks = extractLinks,
   } = config;
 
   const DecoratedLink = (props: LinkProps): ReactElement => (
@@ -35,13 +42,15 @@ export default (config: LinkifyPluginConfig = {}): EditorPlugin => {
       target={target}
       rel={rel}
       component={component}
+      customExtractLinks={customExtractLinks}
     />
   );
 
   return {
     decorators: [
       {
-        strategy: linkStrategy,
+        strategy: (contentBlock, callback) =>
+          linkStrategy(contentBlock, callback, customExtractLinks),
         component: DecoratedLink,
       },
     ],